refatorar módulo de banco de dados para melhorar a inicialização e criar tabelas necessárias
